feat(all): silent posting for archive imports + un-break the Android build
Groundwork for the Instagram importer. An import republishes a whole
back-catalogue at once, so every imported post would have fired a
lock-screen banner at every member of the circle — 900 posts, 900
banners, for content that is usually years old.

Imported posts now go out over the silent path that already existed in
each platform's push leg: the event still delivers, still syncs, and
still reaches offline members through the mailbox — only the banner is
suppressed.

This is a separate authoring entry point on each platform rather than a
`silent` flag on the normal post path or a user-facing toggle, because
silence is a property of importing, not a mode anyone can leave switched
on. No ordinary post can reach it, so nothing real can be published
without notifying the circle. Each takes the original capture time so an
imported archive slots into history by its real date.

  apple    FeedStore.postImported
  android  HavenNet.postImported
  desktop  Engine::post_imported

Also fixes the Android build, which has been broken since 1.4.5 (469):
that release landed a second KeptStoriesStore.all() next to the existing
one, and the ambiguous overload cascaded into MessagesScreen, YouScreen
and StoryLinkScreen — 20+ errors. Both accessors wanted the same stable
snapshot, so they are now one, sorted newest-first. Two nowMs()
subtractions from the same release mixed ULong with Long; both now do
the arithmetic in Long.

Verified: xcodebuild (iOS) BUILD SUCCEEDED, gradlew compileDebugKotlin
BUILD SUCCESSFUL, cargo check clean.

+++ android/app/src/main/java/com/blaineam/haven/ui/MessagesScreen.kt
@@ -710,7 +710,7 @@ private fun Bubble(
     // After 24h, a lone story-shaped visual with no openable story is treated as an expired
     // story reply — never keep showing the resealed media forever.
     val storyLifetimeMs = 24L * 60 * 60 * 1000
-    val replyAge = com.blaineam.haven.core.nowMs() - m.createdAt.toLong()
+    val replyAge = com.blaineam.haven.core.nowMs().toLong() - m.createdAt.toLong()
     val looksLikeLegacyStoryReply = visualRefs.size == 1 && replyAge > storyLifetimeMs
     val showsStoryCard = storyTarget != null
     val showsUnavailable = !showsStoryCard && looksLikeLegacyStoryReply
